move values funcs into values.go, clean up ranges Compare() and Equal()
